+++ tests/testthat/test-datasummary_correlation.R
@@ -1,0 +1,9 @@
+library(modelsummary)
+context('datasummary_correlation')
+
+random_string <- function() {
+  paste(sample(letters, 30, replace=TRUE), collapse="")
+}
+
+
+